+++ src/test/java/ee/sk/mid/ReadmeTest.java
@@ -30,21 +30,15 @@ import static ee.sk.mid.mock.TestData.VALID_SIGNATURE_IN_BASE64;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
+import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
+import java.security.KeyStore;
+import java.security.cert.X509Certificate;
 import java.util.List;
 
-import ee.sk.mid.exception.MidDeliveryException;
-import ee.sk.mid.exception.MidInvalidNationalIdentityNumberException;
-import ee.sk.mid.exception.MidInvalidPhoneNumberException;
-import ee.sk.mid.exception.MidInvalidUserConfigurationException;
-import ee.sk.mid.exception.MidInternalErrorException;
-import ee.sk.mid.exception.MidSessionNotFoundException;
-import ee.sk.mid.exception.MidSessionTimeoutException;
-import ee.sk.mid.exception.MidMissingOrInvalidParameterException;
-import ee.sk.mid.exception.MidNotMidClientException;
-import ee.sk.mid.exception.MidPhoneNotAvailableException;
-import ee.sk.mid.exception.MidUnauthorizedException;
-import ee.sk.mid.exception.MidUserCancellationException;
+import ee.sk.mid.exception.*;
+import ee.sk.mid.integration.MobileIdSSL_IT;
+import ee.sk.mid.mock.TestData;
 import ee.sk.mid.rest.dao.MidSessionStatus;
 import ee.sk.mid.rest.dao.request.MidAuthenticationRequest;
 import ee.sk.mid.rest.dao.request.MidCertificateRequest;
@@ -57,10 +51,13 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.TrustManagerFactory;
+
 /**
  * These tests contain snippets used in Readme.md
  * This is needed to guarantee that tests compile.
- * If anything changes in this class (ecept setUp method) the changes must be reflected in Readme.md
+ * If anything changes in this class (except setUp method) the changes must be reflected in Readme.md
  * These are not real tests!
  */
 public class ReadmeTest {
@@ -74,11 +71,16 @@ public class ReadmeTest {
     MidAuthenticationResult authenticationResult;
 
     @Before
-    public void setUp() {
+    public void setUp() throws Exception {
+        InputStream is = MobileIdSSL_IT.class.getResourceAsStream("/demo_server_trusted_ssl_certs.jks");
+        KeyStore keyStore = KeyStore.getInstance("JKS");
+        keyStore.load(is, "changeit".toCharArray());
+
         client = MidClient.newBuilder()
             .withHostUrl("https://tsp.demo.sk.ee/mid-api")
             .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
             .withRelyingPartyName("DEMO")
+            .withTrustStore(keyStore)
             .build();
 
         MidAuthenticationHashToSign authenticationHash = MidAuthenticationHashToSign.newBuilder()
@@ -93,55 +95,83 @@ public class ReadmeTest {
     }
 
     @Test
-    public void documentConfigureTheClient() {
+    public void documentConfigureTheClient() throws Exception {
+
+        InputStream is = TestData.class.getResourceAsStream("/path/to/truststore.jks");
+        KeyStore trustStore = KeyStore.getInstance("JKS");
+        trustStore.load(is, "changeit".toCharArray());
 
         MidClient client = MidClient.newBuilder()
             .withHostUrl("https://tsp.demo.sk.ee/mid-api")
             .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
             .withRelyingPartyName("DEMO")
+            .withTrustStore(trustStore)
             .build();
     }
 
+    @Test
+    public void documentConfigureTheClientTrustStore() throws Exception {
+
+        InputStream is = TestData.class.getResourceAsStream("/path/to/truststore.jks");
+        KeyStore trustStore = KeyStore.getInstance("JKS");
+        trustStore.load(is, "changeit".toCharArray());
+
+        client = MidClient.newBuilder()
+                .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                .withRelyingPartyName("DEMO")
+                .withHostUrl("https://tsp.demo.sk.ee/mid-api")
+                .withTrustStore(trustStore)
+                .build();
+    }
 
     @Test
+    public void documentConfigureTheClientWithTrustSslContext() throws Exception {
+
+        InputStream is = TestData.class.getResourceAsStream("/path/to/truststore.jks");
+        KeyStore trustStore = KeyStore.getInstance("JKS");
+        trustStore.load(is, "changeit".toCharArray());
+
+        SSLContext trustSslContext = SSLContext.getInstance("TLSv1.2");
+        TrustManagerFactory trustManagerFactory = TrustManagerFactory.getInstance("X509");
+        trustManagerFactory.init(trustStore);
+        trustSslContext.init(null, trustManagerFactory.getTrustManagers(), null);
+
+        client = MidClient.newBuilder()
+                .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                .withRelyingPartyName("DEMO")
+                .withHostUrl("https://tsp.demo.sk.ee/mid-api")
+                .withTrustSslContext(trustSslContext)
+                .build();
+    }
+
+    @Test(expected = MidSslException.class)
+    public void documentConfigureTheClientWithTrustedCertificatesList() throws Exception {
+
+        client = MidClient.newBuilder()
+                .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                .withRelyingPartyName("DEMO")
+                .withHostUrl("https://tsp.demo.sk.ee/mid-api")
+                .withTrustedCertificates("PEM encoded cert 1", "PEM encoded cert 2")
+                .build();
+    }
+
+    @Test(expected = MidMissingOrInvalidParameterException.class)
     public void documentClientWithLongPollingTimeout() {
 
         MidClient client = MidClient.newBuilder()
-            // set hostUrl, hRelyingParty UUID & Name
+            // set hostUrl, relyingPartyUUID, relyingPartyName and trustStore/trustSslContext
             .withLongPollingTimeoutSeconds(60)
             .build();
     }
 
-    @Test
+    @Test(expected = MidMissingOrInvalidParameterException.class)
     public void documentWithPollingSleepTimeoutSeconds() {
 
         MidClient client = MidClient.newBuilder()
-            // set hostUrl, hRelyingParty UUID & Name
+            // set hostUrl, relyingPartyUUID, relyingPartyName and trustStore/trustSslContext
             .withPollingSleepTimeoutSeconds(2)
             .build();
     }
-
-
-    @Test
-    public void documentValidateUserInput() {
-
-
-        try {
-            String nationalIdentityNumber = MidInputUtil.getValidatedNationalIdentityNumber("<national identity number entered by user>");
-            String phoneNumber = MidInputUtil.getValidatedPhoneNumber("<phone number entered by user>");
-        }
-        catch (MidInvalidNationalIdentityNumberException e) {
-            logger.info("User entered invalid national identity number");
-            // display error
-        }
-        catch (MidInvalidPhoneNumberException e) {
-            logger.info("User entered invalid phone number");
-            // display error
-        }
-
-
-    }
-
 
     @Test(expected = MidNotMidClientException.class)
     public void documentRetrieveCert() {
@@ -153,9 +183,8 @@ public class ReadmeTest {
 
         MidCertificateChoiceResponse response = client.getMobileIdConnector().getCertificate(request);
 
-        client.createMobileIdCertificate(response);
+        X509Certificate certificate = client.createMobileIdCertificate(response);
     }
-
 
     @Test
     public void documentCreateFromExistingData() {
@@ -298,6 +327,23 @@ public class ReadmeTest {
         }
         catch (MidInternalErrorException e) {
             logger.warn("MID service returned internal error that cannot be handled locally.");
+        }
+    }
+
+    @Test
+    public void documentValidateUserInput() {
+
+        try {
+            String nationalIdentityNumber = MidInputUtil.getValidatedNationalIdentityNumber("<national identity number entered by user>");
+            String phoneNumber = MidInputUtil.getValidatedPhoneNumber("<phone number entered by user>");
+        }
+        catch (MidInvalidNationalIdentityNumberException e) {
+            logger.info("User entered invalid national identity number");
+            // display error
+        }
+        catch (MidInvalidPhoneNumberException e) {
+            logger.info("User entered invalid phone number");
+            // display error
         }